pyproject.toml
[project]
name = "viggo"
version = "0.1.0"
description = "A spoiler-free lore companion using RAG."
authors = [
    {name = "Marouan"}
]
readme = "README.md"
requires-python = "~3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "pypdf (>=5.6.0,<6.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "groq (>=0.26.0,<0.27.0)",
    "neo4j (>=5.28.1,<6.0.0)",
    "spacy (>=3.8.7,<4.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "mlflow",
    "wandb",
    
    
    
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

